backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
sqlalchemy==2.0.43
pymysql==1.1.2
passlib==1.7.4
python-jose[cryptography]==3.5.0
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.3
numpy==2.1.2
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2
openpyxl==3.1.5
python-multipart==0.0.20
